Update Ookem_Diabetes.md
Added new sources including for the photographs and medical papers, also added a little bit about each machine learning model we used, however I don't understand how to put the images in tabular format does that mean within a table? or to split the images by where they should be in the text?